+++ hackmon.h
@@ -23,7 +23,6 @@
 typedef struct {
 	uint64_t receiveds;
 	uint64_t processeds;
-	uint64_t pidmiss;
 	uint64_t ooms;
 } hackmon_stat_t;
 
@@ -36,10 +35,10 @@ typedef struct {
 	const char *method; /* "ptrace" or "task_for_pid" */
 } process_access_t;
 
-void hackmon_taskforpid(struct timespec *tv, audit_proc_t *, audit_proc_t *)
-     NONNULL(1,2,3);
-void hackmon_ptrace(struct timespec *tv, audit_proc_t *, audit_proc_t *, pid_t)
-     NONNULL(1,2);
+void hackmon_taskforpid(struct timespec *tv, audit_proc_t *, audit_proc_t *,
+                        pid_t) NONNULL(1,2);
+void hackmon_ptrace(struct timespec *tv, audit_proc_t *, audit_proc_t *,
+                    pid_t) NONNULL(1,2);
 
 void hackmon_init(config_t *) NONNULL(1);
 void hackmon_fini(void);